+++ src/dipdup/codegen.py
@@ -27,6 +27,7 @@ from dipdup.config import IndexTemplateConfig
 from dipdup.config import OperationHandlerOriginationPatternConfig
 from dipdup.config import OperationHandlerTransactionPatternConfig
 from dipdup.config import OperationIndexConfig
+from dipdup.config import TokenTransferIndexConfig
 from dipdup.config import TzktDatasourceConfig
 from dipdup.config import default_hooks
 from dipdup.datasources.datasource import Datasource
@@ -101,12 +102,7 @@ class DipDupCodeGenerator:
 
     async def create_package(self) -> None:
         """Create Python package skeleton if not exists"""
-        try:
-            package_path = self._config.package_path
-        except ImportError:
-            self._logger.info('Creating package `%s`', self._config.package)
-            package_path = join(os.getcwd(), self._config.package)
-
+        package_path = self._config.package_path
         touch(join(package_path, '__init__.py'))
 
         models_path = join(package_path, 'models.py')
@@ -116,13 +112,13 @@ class DipDupCodeGenerator:
             write(models_path, models_code)
 
         for subpackage in ('handlers', 'hooks'):
-            subpackage_path = join(self._config.package_path, subpackage)
+            subpackage_path = join(package_path, subpackage)
             touch(join(subpackage_path, '__init__.py'))
 
-        sql_path = join(self._config.package_path, 'sql')
+        sql_path = join(package_path, 'sql')
         touch(join(sql_path, '.keep'))
 
-        graphql_path = join(self._config.package_path, 'graphql')
+        graphql_path = join(package_path, 'graphql')
         touch(join(graphql_path, '.keep'))
 
     async def fetch_schemas(self) -> None:
@@ -215,6 +211,9 @@ class DipDupCodeGenerator:
             elif isinstance(index_config, HeadIndexConfig):
                 pass
 
+            elif isinstance(index_config, TokenTransferIndexConfig):
+                pass
+
             elif isinstance(index_config, IndexTemplateConfig):
                 raise ConfigInitializationException
 
@@ -279,7 +278,7 @@ class DipDupCodeGenerator:
         """Generate handler stubs with typehints from templates if not exist"""
         handler_config: HandlerConfig
         for index_config in self._config.indexes.values():
-            if isinstance(index_config, (OperationIndexConfig, BigMapIndexConfig, HeadIndexConfig)):
+            if isinstance(index_config, (OperationIndexConfig, BigMapIndexConfig, HeadIndexConfig, TokenTransferIndexConfig)):
                 for handler_config in index_config.handlers:
                     await self._generate_callback(handler_config)
 
@@ -289,6 +288,9 @@ class DipDupCodeGenerator:
     async def generate_hooks(self) -> None:
         for hook_configs in self._config.hooks.values(), default_hooks.values():
             for hook_config in hook_configs:
+                # TODO: Skipping deprecated hook, remove in 6.0
+                if hook_config.callback == 'on_rollback':
+                    continue
                 await self._generate_callback(hook_config, sql=True)
 
     async def cleanup(self) -> None:
@@ -374,11 +376,12 @@ class DipDupCodeGenerator:
             code: List[str] = []
             if sql:
                 code.append(f"await ctx.execute_sql('{original_callback}')")
-                if callback == 'on_rollback':
+                if callback == 'on_index_rollback':
                     imports.add('from dipdup.enums import ReindexingReason')
                     code.append('await ctx.reindex(')
                     code.append('    ReindexingReason.rollback,')
-                    code.append('    datasource=datasource.name,')
+                    code.append('    index=index.name,')
+                    code.append('    datasource=index.datasource.name,')
                     code.append('    from_level=from_level,')
                     code.append('    to_level=to_level,')
                     code.append(')')