+++ src/dipdup/enums.py
@@ -5,12 +5,14 @@ class MessageType(Enum):
     operation = 'operation'
     big_map = 'big_map'
     head = 'head'
+    token_transfer = 'token_transfer'
 
 
 class IndexType(Enum):
     operation = 'operation'
     big_map = 'big_map'
     head = 'head'
+    token_transfer = 'token_transfer'
 
 
 class OperationType(Enum):
@@ -47,3 +49,8 @@ class ReindexingAction(Enum):
     exception = 'exception'
     wipe = 'wipe'
     ignore = 'ignore'
+
+
+class TokenStandard(Enum):
+    FA12 = 'fa1.2'
+    FA2 = 'fa2'